nova_aff/requirements.txt
# Core Django
Django==5.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.4.0
drf-yasg==1.21.7

# Database
psycopg2-binary==2.9.9

# Cache & Redis
redis==5.0.1
django-redis==5.4.0

# Environment
python-decouple==3.8
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring & Logging
sentry-sdk[django]==1.40.6

# Image Processing
Pillow==10.1.0

# Utils
requests==2.31.0
celery==5.3.4

# Security
django-ratelimit==4.1.0

# AWS S3 (Optional)
boto3==1.34.0
django-storages==1.14.2 